+++ fe/src/Socket.js
@@ -35,7 +35,8 @@ class Socket {
     let host_is_development =
       (window.location.hostname === "127.0.0.1") |
       (window.location.hostname === "localhost");
-    this.io = io(host_is_development ? "127.0.0.1:8080" : "strate.gg");
+    let options = { withCredentials: true };
+    this.io = io(host_is_development ? "localhost:8080" : "strate.gg", options);
     // Pondering: Can this line ever be slow enough (or light travel fast enough) where the "connect" event happens before it is binded?
     // Answer: very, very, very unlikely. It is both light, and one is traveling (at least) hundreds of miles.
     this.bind_socket_listeners();