+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This file lists all changes to this project, grouped by versions that follow [Se
 ### Added
 
 - Redis to docker-compose for the dev environment
+- Flask-Session to handle server-side sessions with redis as the data backend
 
 ### Changed
 
@@ -15,6 +16,15 @@ This file lists all changes to this project, grouped by versions that follow [Se
 - `Makefile` to use `npm ci` instead of `npm i` to keep package-lock.json consistent across devices
 - These back-end features to use Redis as the datastore:
   - Online user count.
+  - Tracking UIDs (user IDs)
+
+### Fixed
+
+- Issue with socketIO not transferring cookies to development server because
+  Flask-SocketIO could only read cookies that were already sent in a regular HTTP
+  route. Because the Flask backend is only a "websocket API" in the development
+  set-up and not in production, this was causing issues. Fixed by setting the
+  "withCredentials" option for the SocketIO object in the front-end to True.
 
 ## [0.4.1] - 2022-02-22